chore: Set up ESLint and Prettier for root project files
- Configured ESLint and Prettier for Solidity, TypeScript, JSON, and Markdown files in the project root.
- Excluded frontend folder due to its separate existing setup.
- Linted and formatted the relevant files using the new configuration.

+++ contracts/AuctionOffchain.sol
@@ -51,7 +51,7 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
     function submitBid(
         uint16 auctionId
     ) external onlyConfidential onlyInitialized returns (bytes memory) {
-        uint bidAmount = abi.decode(Context.confidentialInputs(), (uint));
+        uint256 bidAmount = abi.decode(Context.confidentialInputs(), (uint256));
         address bidder = msg.sender;
 
         checkBidValidity(auctionId, bidder, bidAmount);
@@ -77,9 +77,7 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
         require(block.timestamp > auction.until, "Auction has not ended");
         require(auction.bids > 0, "No bids");
 
-        (Bid memory winningBid, uint scndBidAmount) = settleVickeryAuction(
-            auctionId
-        );
+        (Bid memory winningBid, uint256 scndBidAmount) = settleVickeryAuction(auctionId);
         AuctionPayout memory payout = AuctionPayout(
             vault,
             auction.payoutAddress,
@@ -105,10 +103,7 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
         uint16 auctionId
     ) external onlyConfidential onlyInitialized returns (bytes memory) {
         Auction storage auction = auctions[auctionId];
-        require(
-            auction.status == AuctionStatus.SETTLED,
-            "Auction is not settled"
-        );
+        require(auction.status == AuctionStatus.SETTLED, "Auction is not settled");
         require(auction.winner == msg.sender, "Only winner can claim token");
         bytes memory tokenBytes = cstore.retrieveToken(auction.tokenDataId);
         bytes memory key = Context.confidentialInputs();
@@ -116,11 +111,7 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
         return Suave.aesEncrypt(key, abi.encode(tokenBytes));
     }
 
-    function checkBidValidity(
-        uint16 auctionId,
-        address bidder,
-        uint bidAmount
-    ) public {
+    function checkBidValidity(uint16 auctionId, address bidder, uint256 bidAmount) public {
         Auction storage auction = auctions[auctionId];
         require(auction.status == AuctionStatus.LIVE, "Auction is not live");
         require(block.timestamp <= auction.until, "Auction has ended");
@@ -136,23 +127,17 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
 
     function hasUserSufficientFunds(
         address user,
-        uint amount,
-        uint auctionEndPlusClaimTime
+        uint256 amount,
+        uint256 auctionEndPlusClaimTime
     ) public returns (bool) {
-        (uint balance, uint64 lockedUntil) = vaultRemote.getBalance(user);
+        (uint256 balance, uint64 lockedUntil) = vaultRemote.getBalance(user);
         return balance >= amount && lockedUntil >= auctionEndPlusClaimTime;
     }
 
-    function signPayout(
-        AuctionPayout memory payout
-    ) internal returns (bytes memory sig) {
+    function signPayout(AuctionPayout memory payout) internal returns (bytes memory sig) {
         string memory pk = cstore.retreivePK(pkDataId);
         bytes32 digest = keccak256(abi.encode(payout));
-        sig = Suave.signMessage(
-            abi.encodePacked(digest),
-            Suave.CryptoSignature.SECP256,
-            pk
-        );
+        sig = Suave.signMessage(abi.encodePacked(digest), Suave.CryptoSignature.SECP256, pk);
     }
 
     function incrementBidCount() internal returns (uint16 bidCount) {
@@ -162,7 +147,7 @@ contract AuctionOffchain is SuaveContract, AuctionCommon {
 
     function settleVickeryAuction(
         uint16 auctionId
-    ) internal returns (Bid memory winningBid, uint scndBestBidAmount) {
+    ) internal returns (Bid memory winningBid, uint256 scndBestBidAmount) {
         Bid[] memory bids = cstore.fetchBids(auctionId);
         (winningBid, scndBestBidAmount) = BidUtils.settleVickeryAuction(bids);
     }
@@ -174,14 +159,10 @@ contract AuctionConfidentialStore {
     string constant BIDCOUNT_NAMESPACE = "auction:v0:bidcount";
     string constant TOKEN_NAMESPACE = "auction:v0:token";
 
-    address[] public genericPeekers = [
-        0xC8df3686b4Afb2BB53e60EAe97EF043FE03Fb829
-    ]; // todo: update after suave update (this exposes storage to everyone)
+    address[] public genericPeekers = [0xC8df3686b4Afb2BB53e60EAe97EF043FE03Fb829]; // todo: update after suave update (this exposes storage to everyone)
 
     function storeBid(Bid memory bid, uint16 auctionId) external {
-        string memory namespace = string(
-            abi.encodePacked(BID_NAMESPACE, auctionId)
-        );
+        string memory namespace = string(abi.encodePacked(BID_NAMESPACE, auctionId));
         address[] memory peekers = new address[](3);
         peekers[0] = address(this);
         peekers[1] = Suave.FETCH_DATA_RECORDS;
@@ -196,19 +177,11 @@ contract AuctionConfidentialStore {
     }
 
     function fetchBids(uint16 auctionId) external returns (Bid[] memory) {
-        string memory namespace = string(
-            abi.encodePacked(BID_NAMESPACE, auctionId)
-        );
-        Suave.DataRecord[] memory dataRecords = Suave.fetchDataRecords(
-            0,
-            namespace
-        );
+        string memory namespace = string(abi.encodePacked(BID_NAMESPACE, auctionId));
+        Suave.DataRecord[] memory dataRecords = Suave.fetchDataRecords(0, namespace);
         Bid[] memory bids = new Bid[](dataRecords.length);
-        for (uint i = 0; i < dataRecords.length; i++) {
-            bytes memory bidBytes = Suave.confidentialRetrieve(
-                dataRecords[i].id,
-                namespace
-            );
+        for (uint256 i = 0; i < dataRecords.length; i++) {
+            bytes memory bidBytes = Suave.confidentialRetrieve(dataRecords[i].id, namespace);
             Bid memory bid = abi.decode(bidBytes, (Bid));
             bids[i] = bid;
         }
@@ -227,38 +200,20 @@ contract AuctionConfidentialStore {
             peekers,
             BIDCOUNT_NAMESPACE
         );
-        Suave.confidentialStore(
-            dataRec.id,
-            BIDCOUNT_NAMESPACE,
-            abi.encode(bidCount)
-        );
+        Suave.confidentialStore(dataRec.id, BIDCOUNT_NAMESPACE, abi.encode(bidCount));
         return dataRec.id;
     }
 
-    function updateBidCount(
-        Suave.DataId bidCountDataId,
-        uint16 bidCount
-    ) external {
-        Suave.confidentialStore(
-            bidCountDataId,
-            BIDCOUNT_NAMESPACE,
-            abi.encode(bidCount)
-        );
+    function updateBidCount(Suave.DataId bidCountDataId, uint16 bidCount) external {
+        Suave.confidentialStore(bidCountDataId, BIDCOUNT_NAMESPACE, abi.encode(bidCount));
     }
 
-    function retrieveBidCount(
-        Suave.DataId bidCountDataId
-    ) public returns (uint16) {
-        bytes memory bidCountBytes = Suave.confidentialRetrieve(
-            bidCountDataId,
-            BIDCOUNT_NAMESPACE
-        );
+    function retrieveBidCount(Suave.DataId bidCountDataId) public returns (uint16) {
+        bytes memory bidCountBytes = Suave.confidentialRetrieve(bidCountDataId, BIDCOUNT_NAMESPACE);
         return abi.decode(bidCountBytes, (uint16));
     }
 
-    function storeToken(
-        bytes memory tokenBytes
-    ) external returns (Suave.DataId) {
+    function storeToken(bytes memory tokenBytes) external returns (Suave.DataId) {
         address[] memory peekers = new address[](3);
         peekers[0] = address(this);
         peekers[1] = Suave.FETCH_DATA_RECORDS;
@@ -273,9 +228,7 @@ contract AuctionConfidentialStore {
         return secretBid.id;
     }
 
-    function retrieveToken(
-        Suave.DataId tokenDataId
-    ) external returns (bytes memory tokenBytes) {
+    function retrieveToken(Suave.DataId tokenDataId) external returns (bytes memory tokenBytes) {
         tokenBytes = Suave.confidentialRetrieve(tokenDataId, TOKEN_NAMESPACE);
     }
 
@@ -294,13 +247,8 @@ contract AuctionConfidentialStore {
         return secretBid.id;
     }
 
-    function retreivePK(
-        Suave.DataId pkDataId
-    ) external returns (string memory) {
-        bytes memory pkBytes = Suave.confidentialRetrieve(
-            pkDataId,
-            PK_NAMESPACE
-        );
+    function retreivePK(Suave.DataId pkDataId) external returns (string memory) {
+        bytes memory pkBytes = Suave.confidentialRetrieve(pkDataId, PK_NAMESPACE);
         return string(pkBytes);
     }
 }
